+++ methods/src/test/kotlin/methods/inequalities/DecimalLinearInequalitiesTest.kt
@@ -42,7 +42,7 @@ class DecimalLinearInequalitiesTest {
                     fromExpr = "3.1 x + 2.2 < 2.9 x - 9.34"
                     toExpr = "0.2 x + 2.2 < -9.34"
                     explanation {
-                        key = methods.solvable.InequalitiesExplanation.MoveVariablesToTheLeftAndSimplify
+                        key = methods.solvable.InequalitiesExplanation.MoveVariablesToTheLeftBySubtraction
                     }
                 }
 
@@ -50,7 +50,7 @@ class DecimalLinearInequalitiesTest {
                     fromExpr = "0.2 x + 2.2 < -9.34"
                     toExpr = "0.2 x < -11.54"
                     explanation {
-                        key = methods.solvable.InequalitiesExplanation.MoveConstantsToTheRightAndSimplify
+                        key = methods.solvable.InequalitiesExplanation.MoveConstantsToTheRightBySubtraction
                     }
                 }
 
@@ -90,7 +90,7 @@ class DecimalLinearInequalitiesTest {
                     fromExpr = "3.1 x + 2.2 >= 1.21"
                     toExpr = "3.1 x >= -0.99"
                     explanation {
-                        key = methods.solvable.InequalitiesExplanation.MoveConstantsToTheRightAndSimplify
+                        key = methods.solvable.InequalitiesExplanation.MoveConstantsToTheRightBySubtraction
                     }
                 }
 
@@ -150,7 +150,7 @@ class DecimalLinearInequalitiesTest {
                     fromExpr = "3.6 x + 2.2 <= 0.4 x + 1.2"
                     toExpr = "3.2 x + 2.2 <= 1.2"
                     explanation {
-                        key = methods.solvable.InequalitiesExplanation.MoveVariablesToTheLeftAndSimplify
+                        key = methods.solvable.InequalitiesExplanation.MoveVariablesToTheLeftBySubtraction
                     }
                 }
 
@@ -158,7 +158,7 @@ class DecimalLinearInequalitiesTest {
                     fromExpr = "3.2 x + 2.2 <= 1.2"
                     toExpr = "3.2 x <= -1"
                     explanation {
-                        key = methods.solvable.InequalitiesExplanation.MoveConstantsToTheRightAndSimplify
+                        key = methods.solvable.InequalitiesExplanation.MoveConstantsToTheRightBySubtraction
                     }
                 }